requirements.txt
pycryptodome
keyring
cryptography
firebase-admin
google-cloud-firestore
customtkinter
pyinstaller
